backend/requirements.txt
django==1.9
django-reversion==2.0.0
django-crispy-forms==1.6.1
django-formtools==1.0
httplib2==0.9.2
xadmin==0.6.1
django-debug-toolbar==1.6
djangorestframework==3.5.1
django-filter==1.0.1
django-extensions==1.7.5
